Refactor import paths for components to use absolute paths for better readability and maintainability
